requirements.txt
Flask==1.1.2
hydra-core==1.3.2
numpy==1.23.5
omegaconf==2.3.0
sentencepiece==0.2.0
setuptools==58.1.0
torch==2.0.0
tqdm==4.64.1
wandb==0.19.8
